Add CampaignUpdated event message
